end of section 12 2nd time

+++ API/Controllers/UsersController.cs
@@ -79,20 +79,38 @@ namespace API.Controllers
         {
             var user = await _userRepository.GetUserByUsernameAsync(User.GetUserName());
 
-            var result = await _photoService.AddPhotAsync(file);
+            var result = await _photoService.AddPhotoAsync(file);
 
             if (result.Error != null) return BadRequest(result.Error.Message);
 
             var photo = new Photo
             {
-                StudentUrl = result.SecureUrl.AbsoluteUri,
-                PublicId = result.PublicId
+                StudentUrl = null,
+                LogoUrl = null
             };
 
-            if (user.Photos.Count == 0)
+            if (user.AppUserType == "ColStudent")
+            {
+                photo.StudentUrl = result.SecureUrl.AbsoluteUri;
+                photo.PublicId = result.PublicId;
+            };
+
+            if (user.AppUserType == "EmpHr")
+            {
+                photo.LogoUrl = result.SecureUrl.AbsoluteUri;
+                photo.PublicId = result.PublicId;
+            };
+
+            if (user.Photos.Count == 0 && user.AppUserType == "ColStudent")
             {
                 photo.IsMain = true;
             }
+
+            if (user.Photos.Count == 0 && user.AppUserType == "EmpHr")
+            {
+                photo.IsMainLogo = true;
+            }
+
             user.Photos.Add(photo);
 
             if (await _userRepository.SaveAllAsync())
@@ -102,7 +120,6 @@ namespace API.Controllers
                 return CreatedAtRoute("GetUser", new { username = user.UserName }, _mapper.Map<PhotoDto>(photo)); // used 3rd overload which 
                 // has the following parameters: string routename, object routeValues, object value
             }
-
 
             return BadRequest("Problem adding photo");
         }
@@ -114,14 +131,30 @@ namespace API.Controllers
 
             var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
 
-            if (photo.IsMain) return BadRequest("This is already your main photo");
-
-            var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
-
-            if (currentMain != null)
+            if (user.AppUserType == "ColStudent")
             {
-                currentMain.IsMain = false;
-                photo.IsMain = true;
+                if (photo.IsMain) return BadRequest("This is already your main photo");
+
+                var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
+
+                if (currentMain != null)
+                {
+                    currentMain.IsMain = false;
+                    photo.IsMain = true;
+                }
+            }
+
+            if (user.AppUserType == "EmpHr")
+            {
+                if (photo.IsMain) return BadRequest("This is already your main logo");
+
+                var currentMain = user.Photos.FirstOrDefault(x => x.IsMainLogo);
+
+                if (currentMain != null)
+                {
+                    currentMain.IsMainLogo = false;
+                    photo.IsMainLogo = true;
+                }
             }
 
             if (await _userRepository.SaveAllAsync()) return NoContent();
@@ -146,6 +179,11 @@ namespace API.Controllers
                 return BadRequest("You cannot delete your main photo");
             }
 
+            if (photo.IsMainLogo)
+            {
+                return BadRequest("You cannot delete your main logo");
+            }
+
             if (photo.PublicId != null)
             {
                 var result = await _photoService.DeletePhotoAsync(photo.PublicId);
@@ -159,7 +197,7 @@ namespace API.Controllers
 
             if (await _userRepository.SaveAllAsync()) return Ok();
 
-            return BadRequest("Failed to delete the photo");
+            return BadRequest("Failed to delete the photo/logo");
         }
     }
 }
